+++ app/src/main/java/gustavo/guterres/leite/tcc/feature/step/StepBuilder.kt
@@ -4,7 +4,7 @@ import gustavo.guterres.leite.tcc.data.entity.model.Step
 
 class StepBuilder {
 
-    fun getFragmentList(steps: List<Step>, onStepFinished: (isRightAnswer: Boolean, points: Double) -> Unit): List<StepFragment> {
+    fun getFragmentList(steps: List<Step>, onStepFinished: (isRightAnswer: Boolean, points: Int) -> Unit): List<StepFragment> {
 
         val fragments = mutableListOf<StepFragment>()
